allow changing ops midway, and store the iterator in a box
Summary:
While using it in some open source projects, I found it to be very useful to be
able to replace the iterator of operations midway.

I don't think perf is really going to be an issue here, so just moving it to
the heap should be fine.
Closes https://github.com/facebookincubator/rust-partial-io/pull/6

Reviewed By: jsgf

Differential Revision: D5143639

Pulled By: sid0

fbshipit-source-id: 31ff9ecda99f624249994fa86f3776a4c870e554