revert D13584553, D13175922, D13166085 to fix rust builds
Summary:
D13166085 breaks Mononoke builds. We got errors

```
rustfmt may have failed to format. See previous 119 errors.
```

These errors happen on formatting thrift files. It looks like the problem is
in the fact that rustfmt for linting is different from rustfmt we are using
while building rust thrift files - https://fburl.com/hn0g95wt. But they seem to
use the same config files, which are incompatible.

For now revert the diff to unblock the builds. Also we have to revert D13584553 and D13175922
because it depends on D13166085.

Reviewed By: lukaspiatkowski

Differential Revision: D13589633

fbshipit-source-id: 78c68068ddcb4c594fd1463e98b48f8f438e46a2