provide edition in .rustfmt.toml
Summary:
When you invoke `rustfmt` tool, it defaults to 2018 edition of formatting.

When RLS is asked to format, it first determines the edition. It will fail to do any formatting if it can't find the edition.

Putting the edition inside fbcode/.rustfmt.toml means that RLS will be willing to go ahead and format any file in fbcode. (unless that file has a more local .rustfmt.toml which omits edition).

Reviewed By: jsgf

Differential Revision: D15425617

fbshipit-source-id: 2cb20778af2383d291d22ba272fc16264741155c