partial-io: add missing `rand` dependency, add quickcheck_types feature and bump major version
Summary:
If one tries to `cargo test` partial-io with all features enabled it will fail, because `rand` is undefined.

In order to turn quickcheck_types module on/off with a single feature a new one had to be created (`quickcheck` cannot be both a meta-feature and a real dependency name). This is a breaking API change so bumping the major version.

Reviewed By: farnz

Differential Revision: D20891362

fbshipit-source-id: ce23d81ee6bf336701b262e174587d3a17e3f6e5